+++ changelog.md
@@ -4,7 +4,8 @@
 
 - Added support for dragging and dropping items between vaults other item piles, vaults, and character sheets
 - Added feature to forbid stacking items in vaults (which can be overridden in individual items)
-- Fixed not being able to right-click to take items from vaults
+- Added option to split stacks of items in the vault item context menu
+- Fixed right-click context menu not working on vaults
 - Minor fixes to the vault grid not interacting well with touch screens
 
 ## Version 2.5.4